+++ lib/src/month_selector/month_year_grid.dart
@@ -26,7 +26,10 @@ class MonthYearGridBuilder extends StatelessWidget {
         (final int index) => MonthButton(
           theme: Theme.of(context),
           date: DateTime(
-              controller.localFirstDate != null ? controller.localFirstDate!.year + page : page, index + 1),
+              controller.localFirstDate != null
+                  ? controller.localFirstDate!.year + page
+                  : page,
+              index + 1),
           localeString: getLocale(context, selectedLocale: controller.locale),
           onMonthSelected: onMonthSelected,
           controller: controller,