+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.0
+- Removed `RxDart` from the code in favor of [Provider](https://pub.dev/packages/provider).
+- Fixed an old error where the year/month becomes zero when changing the orientation [30](https://github.com/hmkrivoj/month_picker_dialog/issues/30).
+
+## 0.8.1
+- Another fix on header arrow [61](https://github.com/hmkrivoj/month_picker_dialog/issues/61).
+
+## 0.8.0
+- Mini fix on header arrow. Bumped to 0.8 because of the braking change on `0.7.1` 
+
+## 0.7.1
+- Breaking: `confirmText` and `cancelText` were replaced by `confirmWidget` and `cancelWidget`. Now it's possible to use any widget instead of only `Text()` in the buttons
+
 ## 0.7.0
 - Added `forceSelectedDate` to fix [41](https://github.com/hmkrivoj/month_picker_dialog/issues/41).The parameter lets you define that the current selected date will be returned if the user clicks outside of the dialog. Needs `dismissible = true`